requirements.txt
flask>=3.1.0
aider-chat>=0.65.1
